backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
numpy>=2.0.0
geopy
scikit-learn
networkx
pydantic
python-dotenv
requests
matplotlib
seaborn
pytest
pytest-asyncio
httpx
alembic
python-multipart
